requirements.txt
asgiref==3.2.10
boto3==1.15.0
boto3-stubs==1.14.60.0
botocore==1.18.0
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
djangorestframework==3.11.1
docutils==0.15.2
ecdsa==0.14.1
idna==2.10
jmespath==0.10.0
opentok==2.10.0
Pillow==7.2.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
python-dateutil==2.8.1
python-http-client==3.3.1
python-jose==3.2.0
pytz==2020.1
requests==2.24.0
rsa==4.6
s3transfer==0.3.3
sendgrid==6.4.7
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
urllib3==1.25.10
uuid==1.30
